clean up async read(): handle case where transport drops
If the reader is stuck blocking on the queue and
transport drops, then the queue will never unblock.
So we drop a sentinel EOF object on the queue when
connection lost is called to unstick any writers.

Change while logic to accept either condition of
transport not closed or queue not empty to avoid
duplication.